+++ apps/explorer/lib/explorer/chain.ex
@@ -3912,13 +3912,6 @@ defmodule Explorer.Chain do
     end
   end
 
-  @spec address_hash_to_smart_contract(Hash.Address.t()) :: SmartContract.t() | nil
-  def address_hash_to_one_smart_contract(hash) do
-    SmartContract
-    |> where([sc], sc.address_hash == ^hash)
-    |> Repo.one()
-  end
-
   @spec address_hash_to_smart_contract_without_twin(Hash.Address.t(), [api?]) :: SmartContract.t() | nil
   def address_hash_to_smart_contract_without_twin(address_hash, options) do
     query =